V0.2 Minimum requirements done, Archive check errors fixed
Archive check compares the required files listed int the  WAEX.index to the files in the directory. If  a required file is missing the counter is incremented and the missing file is noted in the exported error.csv file. Te errors.csv lists the po# and the files that it is missing. 

TODO:
 Have the csv information show up in a table model so the user can click on a specific po# which redirects them to the "create file" page