Fixed Archive Check Table
There was an issue where the PO# would be displayed when there was no missing files. The root of the file is that an empty vector was getting sent to the table model which was producing the empty line. I also added "errorData.erase(errorData.end()-1);" to the archive check this gets rid of the PO# which was producing the empty vector. To fix this issue with the empty vector i just needed to check the length of the vector before loading it into the table. If it is empty i don't load it into the table model.